requirements/local.txt
-r ./base.txt

ipdb==0.13.9

django-extensions==3.1.5

mypy==0.910
mypy-extensions==0.4.3
pytest==6.2.5
pytest-django==4.5.2
pytest-sugar==0.9.4
python-dateutil==2.8.2

Faker==10.0.0
flake8==4.0.1
